+++ package.yml
@@ -1,0 +1,11 @@
+# This file represents the root package of the application
+# Please validate the configuration using `packwerk validate` (for Rails applications) or running the auto generated
+# test case (for non-Rails projects). You can then use `packwerk check` to check your code.
+
+# Turn on dependency checks for this package
+enforce_dependencies: true
+
+# A list of this package's dependencies
+# Note that packages in this list require their own `package.yml` file
+dependencies:
+  - "app/packages/oauth"
