+++ packs/oauth/package.yml
@@ -1,0 +1,4 @@
+enforce_dependencies: true
+enforce_privacy: false
+dependencies:
+  - .
