+++ packs/oauth/services/oauth/user_manager_service.rb
