+++ packs/oauth/spec/request/omniauth_callbacks_request_spec.rb
